chore(pages/discounts/): reorganizando árvore de arquivos para discontos
